requirements.txt
langchain
ipykernel
python-dotenv
langchain_community 
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma==0.1.1
langchain_groq
langserve==0.2.2
fastapi
uvicorn2